Refactor project structure and enhance component imports
- Updated import paths for ProjectTasks and related components to reflect new directory structure.
- Changed input type from text to search in InviteModal for better user experience.
- Improved styling consistency in InviteModal with dark mode adjustments.
- Removed unused AddProjectTask, ProjectTasks, ProjectTasksDetails, TaskCard, and TaskColumn components to streamline the codebase.
- Updated AllTasks component to reflect changes in ProjectTasksDetails import and handling of task details modal.
