+++ packages/Frontend/src/component/pages/Projects/projectTasks/TaskCard.jsx
@@ -2,11 +2,13 @@ import { Calendar, SquarePen, Trash2 } from "lucide-react";
 
 const TaskCard = ({ task, onDelete, onEdit, onClick }) => {
   return (
-    <div className="bg-white border dark:bg-[#2D2D2D] dark:border-0 border-gray-200 rounded-[12px] p-3 shadow-sm max-w-[350px]">
+    <div
+      className="bg-white border dark:bg-[#2D2D2D] dark:border-0 border-gray-200 rounded-[12px] p-3 shadow-sm max-w-[350px] cursor-pointer"
+      onClick={() => onClick(task)}
+    >
       <div className="flex justify-between items-start mb-3 border-b border-gray-200 pb-2">
         <h3
-          className="text-sm dark:text-gray-200 font-medium truncate cursor-pointer"
-          onClick={() => onClick(task)}
+          className="text-sm dark:text-gray-200 font-medium truncate"
         >
           {task.title?.split(" ").length > 5
             ? task.title.split(" ").slice(0, 5).join(" ") + "..."
@@ -14,13 +16,19 @@ const TaskCard = ({ task, onDelete, onEdit, onClick }) => {
         </h3>
         <div className="flex">
           <button
-            onClick={() => onDelete(task._id)}
+            onClick={(e) => {
+              e.stopPropagation();
+              onDelete(task._id);
+            }}
             className="text-red-400 hover:text-red-600"
           >
             <Trash2 width={19} height={24} />
           </button>
           <button
-            onClick={() => onEdit(task)}
+            onClick={(e) => {
+              e.stopPropagation();
+              onEdit(task);
+            }}
             className="text-gray-400 hover:text-gray-600 ml-2"
           >
             <SquarePen width={19} height={24} />
